+++ 15903.cpp
@@ -5,26 +5,26 @@
 using namespace std;
 
 int main() {
-	ios::sync_with_stdio(false); // 15903
+	ios::sync_with_stdio(false);
 	cin.tie(NULL);
 	cout.tie(NULL);
 
-	int N, M;
+	long long N, M;
 	cin >> N >> M; 
 
-	int a;
-	int answer = 0;
+	long long a;
+	long long answer = 0;
 
 	priority_queue<long long, vector<long long>, greater<long long>> pq;
 
-	for (int i = 0; i < N; ++i) {
+	for (long long i = 0; i < N; ++i) {
 		cin >> a;
 		pq.push(a);
 	}
 
-	for (int i = 0; i < M; ++i) {
-		int tmp = 0;
-		for (int j = 0; j < 2; ++j) {
+	for (long long i = 0; i < M; ++i) {
+		long long tmp = 0;
+		for (long long j = 0; j < 2; ++j) {
 			tmp += pq.top();
 			pq.pop();
 		}
@@ -32,7 +32,7 @@ int main() {
 		pq.push(tmp);
 	}
 	
-	for (int i = 0; i < N; ++i) {
+	for (long long i = 0; i < N; ++i) {
 		answer += pq.top();
 		pq.pop();
 	}
